Export wrong or skipped formulas to file.
This is the default behaviour, and the file is saved under R
working directory, named 'cuoti.csv'. The formulas will be
appended to the file each time, so the previous formulas
won't be erased.

Minor changes:
1. Improve the calculation of correctness, ignore skipped
formulas.
2. Change the meaning of 'length', '1' means numbers between
0 and 20, '2' means numbers between 10 and 50, if 'length' is
ignored, then numbers are between 0 and 50.
3. Add the 4th column '正误' in the result data frame, which
makes the wrong results be identified easier.